Update code on flems.io and update the link to it
This page contains a link to an example running on **flems.io**
The code on **flems.io** depends on `mithrill@1.1.6` and doesn't work with `mithril@2.0.0-rc.4`.
I added `mithril@2.0.0-rc.4` on **flems.io** as a dependancy and fixed some issues caused by `mithrill.withAttr` no longer existing.
Now code on this page and on **flems.io** are identical.